+++ src/Makefile.am
@@ -148,6 +148,7 @@ BITCOIN_CORE_H = \
   globals.h \
   httprpc.h \
   httpserver.h \
+  index/base.h \
   index/txindex.h \
   indirectmap.h \
   init.h \
@@ -267,6 +268,7 @@ libbitcoin_server_a_SOURCES = \
   globals.cpp \
   httprpc.cpp \
   httpserver.cpp \
+  index/base.cpp \
   index/txindex.cpp \
   init.cpp \
   interfaces/handler.cpp \
